Abstract out display filtering logic into useFilteredDisplays
