+++ src/screens/home/displays/DisplayGrid.tsx
@@ -1,42 +1,27 @@
 import { Divider } from '@heroui/react';
 import { DisplayCard } from '@luna/components/DisplayCard';
 import { displayLayoutId } from '@luna/constants/LayoutId';
-import { Users } from '@luna/contexts/api/model/ModelContext';
-import { DisplayPin, usePinnedDisplays } from '@luna/hooks/usePinnedDisplays';
+import { useFilteredDisplays } from '@luna/hooks/useFilteredDisplays';
+import { DisplayPin } from '@luna/hooks/usePinnedDisplays';
 import { motion } from 'framer-motion';
-import { useMemo } from 'react';
 import { InView } from 'react-intersection-observer';
 import { Link } from 'react-router-dom';
 
 export interface DisplayGridProps {
-  users: Users;
   searchQuery: string;
   displayWidth: number;
 }
 
-export function DisplayGrid({
-  users,
-  searchQuery,
-  displayWidth,
-}: DisplayGridProps) {
-  const pinnedDisplays = usePinnedDisplays();
-
-  // Filter the models case-insensitively by the search query
-  const filteredUsers = useMemo(
-    () =>
-      [...users.all.keys()].filter(
-        username =>
-          !pinnedDisplays.has(username) &&
-          username.toLowerCase().includes(searchQuery.toLowerCase())
-      ),
-    [pinnedDisplays, searchQuery, users.all]
-  );
+export function DisplayGrid({ searchQuery, displayWidth }: DisplayGridProps) {
+  const { pinnedDisplays, filteredUsernames } = useFilteredDisplays({
+    searchQuery,
+  });
 
   // Disable animations automatically if there are too many displays for
   // performance reasons. Unfortunately we can't seem to change the layoutId
   // after the component has mounted, so even if a user filters down the view
   // the displays might not animate: https://github.com/framer/motion/issues/2075
-  const animationsEnabled = filteredUsers.length <= 360;
+  const animationsEnabled = filteredUsernames.length <= 360;
 
   return (
     <div className="flex flex-wrap gap-4 justify-center">
@@ -50,16 +35,14 @@ export function DisplayGrid({
         />
       ))}
       <Divider />
-      {[...filteredUsers]
-        .sort((u1, u2) => u1.localeCompare(u2))
-        .map(username => (
-          <DisplayLink
-            key={username}
-            username={username}
-            animationsEnabled={animationsEnabled}
-            displayWidth={displayWidth}
-          />
-        ))}
+      {filteredUsernames.map(username => (
+        <DisplayLink
+          key={username}
+          username={username}
+          animationsEnabled={animationsEnabled}
+          displayWidth={displayWidth}
+        />
+      ))}
     </div>
   );
 }
